WEBNEW-211 :truck: Moved render methods to src/components
